+++ lib/medea_flutter_webrtc.dart
@@ -1,25 +1,25 @@
 library medea_flutter_webrtc;
 
 export 'src/api/devices.dart'
-    if (dart.library.html) 'src/platform/web/fake_media.dart';
-export 'src/api/peer.dart' if (dart.library.html) 'none.dart';
-export 'src/api/sender.dart' if (dart.library.html) 'none.dart';
-export 'src/api/transceiver.dart' if (dart.library.html) 'none.dart';
-export 'src/api/parameters.dart' if (dart.library.html) 'none.dart';
+    if (dart.library.js_interop) 'src/platform/web/fake_media.dart';
+export 'src/api/peer.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/api/sender.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/api/transceiver.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/api/parameters.dart' if (dart.library.js_interop) 'none.dart';
 export 'src/api/send_encoding_parameters.dart'
-    if (dart.library.html) 'none.dart';
-export 'src/model/capability.dart' if (dart.library.html) 'none.dart';
-export 'src/model/constraints.dart' if (dart.library.html) 'none.dart';
+    if (dart.library.js_interop) 'none.dart';
+export 'src/model/capability.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/model/constraints.dart' if (dart.library.js_interop) 'none.dart';
 export 'src/model/constraints.dart' show FacingMode;
-export 'src/model/device.dart' if (dart.library.html) 'none.dart';
-export 'src/model/ice.dart' if (dart.library.html) 'none.dart';
-export 'src/model/peer.dart' if (dart.library.html) 'none.dart';
-export 'src/model/sdp.dart' if (dart.library.html) 'none.dart';
-export 'src/model/track.dart' if (dart.library.html) 'none.dart';
-export 'src/model/stats.dart' if (dart.library.html) 'none.dart';
-export 'src/model/transceiver.dart' if (dart.library.html) 'none.dart';
+export 'src/model/device.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/model/ice.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/model/peer.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/model/sdp.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/model/track.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/model/stats.dart' if (dart.library.js_interop) 'none.dart';
+export 'src/model/transceiver.dart' if (dart.library.js_interop) 'none.dart';
 export 'src/platform/audio_renderer.dart';
 export 'src/platform/native/video_view.dart'
-    if (dart.library.html) 'src/platform/web/video_view.dart';
+    if (dart.library.js_interop) 'src/platform/web/video_view.dart';
 export 'src/platform/track.dart';
 export 'src/platform/video_renderer.dart';