+++ lib/src/platform/track.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:medea_flutter_webrtc/src/model/constraints.dart';
 import '/src/model/track.dart';
 
@@ -97,10 +99,10 @@ abstract class MediaStreamTrack {
   /// Returns [width] of this [MediaStreamTrack].
   ///
   /// [width]: https://w3.org/TR/mediacapture-streams#dfn-width
-  Future<int?> width();
+  FutureOr<int?> width();
 
   /// Returns [height] of this [MediaStreamTrack].
   ///
   /// [height]: https://w3.org/TR/mediacapture-streams#dfn-height
-  Future<int?> height();
+  FutureOr<int?> height();
 }