+++ lib/src/platform/video_renderer.dart
@@ -3,10 +3,10 @@ import 'package:flutter/foundation.dart';
 import 'track.dart';
 
 import '/src/platform/native/video_renderer.dart'
-    if (dart.library.html) 'web/video_renderer.dart';
+    if (dart.library.js_interop) 'web/video_renderer.dart';
 
 export 'native/video_renderer.dart'
-    if (dart.library.html) 'web/video_renderer.dart';
+    if (dart.library.js_interop) 'web/video_renderer.dart';
 
 @immutable
 class RTCVideoValue {