+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All user visible changes to this project will be documented in this file. This p
 
 - Upgraded [libwebrtc] to [126.0.6478.182-r2] version. ([#155], [#162])
 - Disable [H264] software encoders and decoders. ([#153])
+- Migrated from [`dart:html`] to [`package:web`]. ([#164])
 
 ### Fixed
 
@@ -45,6 +46,9 @@ All user visible changes to this project will be documented in this file. This p
 [#155]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/155
 [#156]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/156
 [#162]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/162
+[#164]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/164
+[`dart:html`]: https://dart.dev/libraries/dart-html
+[`package:web`]: https://pub.dev/packages/web
 [126.0.6478.182-r2]: https://github.com/instrumentisto/libwebrtc-bin/releases/tag/126.0.6478.182-r2